adaptive ddmrg bug fix:
a subtle bug was lurking in the high level loop
that calls evolve as many times as there are
operators to apply to get the first Lanczos
vector for the spectral function to compute.
That loop was being shortened to only one iteration
if no stage was disabled. This was a bug; it
should have been shortened only if all stages
were converging. In the future this won't matter
since the multiple evolve won't be called by the
adaptive ddmrg, as it is no longer needed
after all stages are converging (i.e. all operators
have been applied).