GeometryUtils added and restriction to Range class
If a loop over a range is created, it might be
possible that not all processors visit the
interior of the loop. Therefore, calls to mpi_reduce or
any other group communication will fail (or deadlock)
To avoid this, Range now enforces the restriction that
nprocs must divide the size of the range.