threading input checks, fixed wft+su(2) issue, but see below
When su(2) is enabled, vectors are still dropped on the left corner
in some instances. Results will still be correct if targeting the
ground state only. A minimal performance impact at that corner is
expected. If targeting anything that requires continued transport
of states with the wft (like tst, dynamic dmrg, correction vector, etc)
results will be wrong if a vector is ever dropped.