new SolverOptions noSaveWft noSaveStacks noSaveData minimizeDisk
If noSaveWft is in SolverOptions then Wft data is not written
If noSaveStacks is in SolverOptions then stack data is not written
If noSaveData is in SolverOptions then basis and vector data
is not written.

minimizeDisk is equivalent to noSaveWft,noSaveStacks,noSaveData

In all cases files are overwritten with minimal data.