KroneckerDumper feature added
KroneckerDumper will print matrices and
related info to create the stored H matrix.

To enable KroneckerDumper add KroneckerDumper
to solver options in the input file.

KroneckerDumper has an internal counter that
gets incremented at each call of the constructor
of ModelHelperLocal in Diagonalization.h.
KroneckerDumperBegin=
is an optional label in the input file that
disables KroneckerDumper if internal counter
is less than the value given.
KroneckerDumperEnd=
is an optional label in the input file, that
disables KroneckerDumper if internal counter
is greater or equal to the value given.

KroneckerDumper should only be used with
MatrixVectorStored as a SolverOption.
KroneckerDumper should only be used with
GroundStateTargeting.
KroneckerDumper should not be used with
SU(2) symmetry enabled.
KroneckerDumper affects only the ModelHelperLocal
object in Diagonalization.h.