Parallelize H_L and H_R with connections
H*y = H_L*y + H_R*y + connections*y.
Now, H_L*y and H_R*y are parallelized within
connection, so that a maximum of c + 2 pthreads are useful,
where c is the number of connections.

The theoretical maximum number of pthreads that are useful
is equal to C + 2, where
C = number of connection = 2*G*M
where G = geometry factor
and M = model factor
G = 1 for a chain
G = leg for a ladder with leg legs. (For instance, G=2 for a 2-leg ladder).
M = 2 for the Hubbard model

In general, M = \sum_{term=0}^{terms} dof(term)
where terms and dof(term) is model dependent.
To find M for a model, go to the Model directory and see LinkProduct*.h file.
the return of function terms() for terms.
For dof(term) see function dof(SizeType term,...).
For example, for the HubbardModelOneBand, one must look at
src/Model/HubbardOneBand/LinkProductHubbardOneBand.h
In that file, terms() returns 1, so that terms=1
Therefore there is only one term: term = 0.
And dof(0,...) = 2, as you can see in  LinkProductHubbardOneBand.h.
Then M = 2 in this case.