added BalancerDefault and NoPthreadsNg
There are two interfaces to [No]Pthreads[Ng]:
the old one, and the new one with Ng.
For now, the old one is the default.
The new one can be enabled with
in your driver before you include
Parallelizer.h
After a period of testing, the new (Ng) interface
will become the default