Note about Frequency and Time Observables
Frequency observables:
|correction vector> = 1/(w-H+i eta) A |gs>
observable = <correction vector| B |gs>
Because the correction vector is taken adjoint,
the observable changes the sign of eta

Time observables
|time vector> = exp(-iHt) A |gs>
observable = <time vector| B |gs>
Because the time vector is taken adjoint,
the observable changes the sign of time t.

Moreover, note that (A^\dagger\times B) must
conserve symmetry to get non-zero observable.
Therefore, if one puts (say) a creation operator
in the input file, then a creation operator must also be
put in the in-situ measurement to obtain a possible non-zero.