DenseSparseThreshold label added
DenseSparseThreshold=threshold
is an optional standlone label in the input file
that controls when a possible sparse matrix is
considered dense. threshold is a real number in [0,1]
If threshold = 0 then all possible sparse matrices
are considered dense.
If threshold = 1 then no sparse matrix is considered dense.
This is currently used only by MatrixDenseOrSparse, which
in turn is currently used only by MatrixVectorKron.