operator specification in the command line
(1) :filename
where filename must be a non-empty string. Every ocurrence
of the character $ in filename is replaced by the current site.
The resulting string names a file that must exist and be
readable by DMRG++ process owner. The file must contain an
operator in file form. See operator specification in file form.

(2) name[site]?dof'
Name is mandatory; it is a non-empty string [a-zA-Z_]+
Optionally, a site may be specified by enclosing it in brakets.
Optionally, the character ? may be present, and if so, it must
be followed by a non-negative integer [0-9]+ that becomes dof.
Optionally, a single quote ' may be present at the end, to mean
that the whole operator must be transposed.
The operator is obtained by calling the model member function
naturalOperator(name,site,dof) and optionally tranposing it.
Note that the fermion sign and the SU(2) properties of the
resulting operator are determined by the member function in
the corresponding model class.