recover feature: user interface
recoveryEnableRead is needed in SolverOptions to enable
recovery if previous run crashed.
RecoverySave=something
is needed so that files are saved for recovery. Files
are deleted if run finishes successfully.
something can be "no" which is the default and disables writing.
something can be l%1 which saves every finite loop or
l%2 which saves all even loops, etc.
something can be dt>600 which saves every 600 seconds (10 minutes),
with a granularity equal to a one-site movement.