LanczosVectors performance
If lotaMemory == false nothing has changed, and
what follows assumes lotaMemory has been set to true
LanczosVectors now preallocates a matrix of
rows = size of space
cols = number of Lanczos steps requested
and saves the Lanczos vectors as the steps are performed
If there is an early exit and not all Lanczos steps are
performed, LanczosVectors then resizes the array to the
actually needed size.
This change avoids the expensive copy of all Lanczos vectors
to storage at the end of the Lanczos, that was done in the
previous code.
It should, on the other hand, consume more RAM depending on early
exit and maximum number of steps chosen