KronUseLowerPart: temporary fix for complex numbers
KronUseLowerPart must transpose conjugate and not just
transpose when calling kronMult in KronConnections.h
If FIX_KRON_USE_LOWER_PART is defined in the preprocessor
then a temporary measure will take place:
Copies of A and B are made, then conjugated, and used.
This is wasteful of RAM and instead the fix must occur
within KronUtil.