restart and InfiniteLoopKeptStates
To restart from a previous well finished run,
(1) Add the option restart to SolverOptions and

(2) Add the hdf5 file to restart from in
InfiniteLoopKeptStates=
or in
RestartFilename=

If using InfiniteLoopKeptStates=filename and the filename is an integer
then surround it with double quotes to have it interpreted as a string.

-----

A FATAL ERROR occurs in the following circumstances
(a) restart in SolverOptions, and both
InfiniteLoopKeptStates= not an integer, and RestartFilename= given

(b) restart in SolverOptions and
InfiniteLoopKeptStates= is an integer, and RestartFilename= not given

(c) no restart in SolverOptions and
InfiniteLoopKeptStates= not an integer

(d) no restart in SolverOptions and
RestartFilename= given

----

A WARNING is printed in the following circumstances

(e) restart in SolverOptions, and RestartFilename= given, and
InfiniteLoopKeptStates= given with an integer value.
Then the integer given in InfiniteLoopKeptStates= is ignored.