SolverOptions=wftAccelPatches
This option forces wftAccelPatches for twositedmrg

Here's the current algorithm:
If no WftAccel* option is present then
WftAccelPatches is used for onesitedmrg, and
WftAccelBlocks for twositedmrg
If wftNoAccel option is present then acceleration isn't used.
If wftAccelPatches option is present then wftAccelPatches
is used for both onesitedmrg and twositedmrg.
DO NOT use more than one WftAccel* option. If you do, what
happens will be officially UNDEFINED, and you will NOT be
warned!