Restart new input lines
RestartSourceTvForPsi=integer
Optional. If present and non-negative, set the g.s. of this targeting to the
previous run target vector indicated by this number.

RestartMappingTvs (vector of integers)
Optional. If present the number of numbers to follow must be provided first,
as for all vectors in InputNg legacy, and with the appropriate syntax for Ainur.
Sets the i-th target vector for this targeting to the previous run target
vector number MappingTvs[i], if MappingTvs[i] is non-negative; skips it if negative.