RecoverySave
This is a comma-separated list options; whitespace isn't allowed.
Supported options are as follows.

l%n, where n is a positive integer, and indicates that save occurs when
the loop index is divisible by n. So, l%1 saves every loop.
CAUTION: Loops that don't end at the lattice border shouldn't be saved, because
the minimum saving unit is a full sweep.

no, which disables RecoverySave, and is assumed if RecoverySave= is
absent from the input file. In other words, RecoverySave is disabled by default.

keep, which keeps recovery files even if the run finishes normally.
By default, all recovery files are deleted if the run finishes normally.

M=n, where n is the maximum number of recovery files that will be saved, before
the oldest file is overwritten. Defaults to 10.