Braket Actions
Observe code for 4-point accepts braket actions.
Example,
./observe -f "{action=something},<gs|c';c;c?1';c?1|gs>"

Actions affect all 4-point brakets after the action is given.
To change actions, write them before the next 4-point.
For example, in
./observe -f "{action=a},<...>,<...>,{action=b},<...>"
action=a applies to the first 2 brakets, and action=b
to the last braket.

Actions are written as
action=predicate awesome
Where the predicate awesome accepts the 4 points
as %0, %1, %2, and %3