Sectors and Excited for NGSTs
Whenever a NGST calls AOE::getPhi
the AOE class does
|phinew> = operators |phi old>

It used to be that |phi old> was taken
to be the ground state.
But now the ground state can be specified
in sectors and excited, so there is
ambiguity as to what is |phi old>
This can now be specified as
TSPApplyTo="|Q0X0>";
or indices for sector and excited instead
of 0 and 0 as shown in this example.