Defining custom operators in input file
DefineOperators="def0,def1,...";
can now be added to the command line.
This is a comma separated list of definitions, each of the form
name:opspec
where name is the name of the operator you want to define and
must be in [a-zA-Z0-9_]+ and must start with a letter not
a number or underscore; opspec is an operator specification.
You may NOT use filename specification for the opspec.

At most one DefineOperators= line may appear in the input.
Operators may not be NOT multiply defined.
Operators so defined may NOT be made trackable.
For non-Ainur inputs, omit the double quotes and the semicolon.
The semantics of the operators so defined are model-dependent.
Omit the comma if only one definition is desired.

Example.

Model="FeAsBasedSc";
FeAsMode="INT_PAPER33";
Orbitals=3;
DefineOperators="na:n?0+n?3,nb:n?1+n?4,nc:n?2+n?5";

defines 3 operators, na = n_{a, up} + n_{a, down},
and so on for the other orbitals.
(Remember that orbitals change first, then spin.)

Now you can run with
./dmrg -f input.ain "<gs|na|gs>,<gs|nb|gs>,<gs|nc|gs>"
and use them in the observe code
./observe -f observe "<gs|na;nb|gs>"