restart NGST --> different NGST
For a restart NGST --> NGST, the
names of prev and this NGST are compared
to LEGACY and if neither is LEGACY
then the NGST --> NGST is considered
SAME if names are equal.
If either is LEGACY then sizes
of prev and this tvs are compared, and
the NGSTs are considered SAME if sizes
are equal.
If NGSTs are NOT same then time step
and floating point time are set to zero.