KondoEx: bug fix when using just one symmetry
When KondoEx is using just one symmetry
that symmetry is TargetSzPlusConst and stored in other[0];
in order to conserve electronic parity we need to
indicate to the Engine that other[0] is NOT the number
of electrons, so that we need
ifPresentOther0IsElectrons = false;
in this case. This commit fixes this bug.