Initial Bath Params for the Fit
If InitBathRa= or InitBathRb= are given in
the input file, then its values are used
to set the initial Vs and Es values with the formula
results[i] = InitBathRa*rng() + InitBathRb

If a vector is given with InitBathVector
in the input file, then the vector is used
to set the initial Vs and Es values with the formula
results[i] = vector given [i]

If InitBathRa= or InitBathRb= are specified
then InitBathVector cannot be specified,
and viceversa.

The default behavior is to set InitBathRa=1.

Moreover, the input line InitBathReset
can be set to 1 to indicate that
the initial bath parameters need to
be set at every Fit attempt.

The default value of InitBathReset
is 0 (false) if unset.