OpOnSiteThreshold= line added
OpOnSiteThreshold=x
is an optional label.
It means that an operator isn't rotated if
its on-site index is greater or equal to x.

The on-site index of an operator is the index it
has within a given site. For example, for the
one-band Hubbard model cup has on-site index 0,
and cdown has on-site index 1.

This feature is ignored if set to 0, which is
also the default.
This feature is also ignored if the model has a
site-dependent Hilbert space (SDHS).