scripts: inSituVsOmega.pl
inSituVsOmega.pl runForinput3.cout label 42 1

If the first argument is an actual filename (like in the example above)
ending in number.cout, then number is considered the start;
In the example above, 3 is the start number.
If a root like runForinput is given instead, then 0 is
considered the start number.

In the example above, the last two arguments, 42 and 1, are optional.

If a third argument is given (like 42 in the example above)
then that is a total, and the last .cout file to be considered.
Moreover, if a total is given then in-between files that
do not exist are skipped, instead of the script dying.

If a fourth argument is given and it's non-zero
(like 1 in the example above) then the max difference among all
occurrences of the value in label is printed in the last column