scripts: lorentzian.pl
lorentzian.pl filename omegaBegin omegaTotal omegaStep delta norm

filename is a two column formatted file, where
the first column contains weights and the second
column contains the omegas for each weight.

The loop is from i=0 to i<omegaTotal with
omega = omegaBegin +  omegaStep*i

The Lorentzian has weight delta.

The norm is optional and, if provided, the function
will be normalized so that the sum over omegas
equals norm.