termGeomReplacement
This commit fixes a bug in the implementation of
termGeometryReplacements in ModelLinks.

This commit also reduces effective terms for
(i) KitaevExtended from six to five, by aliasing
sysx to sxsy, and
(ii) KitaevWithGammas from nine to six, by aliasing
sysz to szsy
sxsz to szsx
sysx to sxsy, and
(iii) KitaevExtendedWithCharge from seven to six, and
(iv) KitaevWithGammasWithCharge from ten to seven.