IoNgSerializer: write std::vector<std::complex<T> >
writing std::vector<std::complex<T> > in IoNgSerializer
now writes a marker before, with name equal to
name of vector + "ComplexOrReal" and content "C"
This marker isn't written if the vector is real.
readInternal tries to read the marker allowing
reads from real --> complex, in addition to the
real --> real, and complex --> complex.